requirements.txt
py==1.4.29
pytest==3.0.2
requests==2.20.0
wheel==0.24.0
Pillow
pandas